+++ backends/slack/src/main/kotlin/com/gatehill/corebot/backend/slack/model/GroupsCreateResponse.kt
@@ -1,7 +1,0 @@
-package com.gatehill.corebot.backend.slack.model
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class GroupsCreateResponse(val ok: Boolean,
-                                val group: SlackGroup)