+++ backends/slack/src/main/kotlin/com/gatehill/corebot/backend/slack/model/SlackGroup.kt
@@ -1,8 +1,0 @@
-package com.gatehill.corebot.backend.slack.model
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class SlackGroup(val id: String,
-                      val name: String,
-                      val members: List<String>)