+++ backends/slack/src/main/kotlin/com/gatehill/corebot/backend/slack/model/UsersListResponse.kt
@@ -1,7 +1,0 @@
-package com.gatehill.corebot.backend.slack.model
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class UsersListResponse(val ok: Boolean,
-                             val members: List<SlackUser>)