+++ scripts/docker-build.sh
@@ -9,7 +9,6 @@ IMAGES=(
     "generic"
     "slack-deploy"
     "slack-items"
-    "http-slack"
 )
 
 while getopts "e" OPT; do
@@ -26,7 +25,6 @@ function getImageSuffixes() { case $1 in
     generic) echo "corebot-generic" ;;
     slack-deploy) echo "corebot corebot-slack-deploy" ;;
     slack-items) echo "corebot-slack-items" ;;
-    http-slack) echo "corebot-http-slack" ;;
 esac }
 
 function buildImage()